fix: factor out transport error
