+++ src/error.rs
@@ -1,4 +1,60 @@
-use std::{error::Error as StdError, fmt};
+use crate::Error as TransportError;
+use std::ops::ControlFlow;
+use std::{error::Error as StdError, fmt, io};
+use tracing::debug;
+
+/// Handles errors that occur during connection acceptance.
+///
+/// This function determines whether an error should be treated as fatal (breaking the accept loop)
+/// or non-fatal (allowing the loop to continue). This handler is crucial for maintaining the
+/// stability and reliability of the server by appropriately handling different types of
+/// errors at the accept stage.
+///
+/// # Arguments
+///
+/// * `e` - The error to handle, which can be converted into the crate's [`crate::Error`] type.
+///
+/// # Returns
+///
+/// * [`ControlFlow::Continue(())`] if the error is non-fatal and the accept loop should continue.
+/// * [`ControlFlow::Break(Error)`] if the error is fatal and the accept loop should terminate.
+///
+/// # Error Handling
+///
+/// The function categorizes errors as follows:
+/// - Non-fatal errors:
+///       [`io::ErrorKind::ConnectionAborted`],
+///       [`io::ErrorKind::Interrupted`],
+///       [`io::ErrorKind::InvalidData`],
+///       [`io::ErrorKind::WouldBlock`]
+/// - Fatal errors: All other error types
+pub(crate) fn handle_accept_error(e: impl Into<TransportError>) -> ControlFlow<TransportError> {
+    // Convert the input error into our crate's Error type
+    let e = e.into();
+
+    // Log the error for debugging purposes
+    debug!(error = %e, "connection accept loop error");
+
+    // Check if the error is an I/O error
+    if let Some(e) = e.downcast_ref::<io::Error>() {
+        // Determine if the error is non-fatal
+        if matches!(
+            e.kind(),
+            io::ErrorKind::ConnectionAborted  // Connection was aborted by the client or network
+                | io::ErrorKind::Interrupted  // The operation was interrupted (e.g., by a signal)
+                // Raised if TLS handshake failed
+                | io::ErrorKind::InvalidData  // Received invalid data (might be temporary)
+                | io::ErrorKind::WouldBlock // Operation would block (common in non-blocking I/O)
+        ) {
+            // For non-fatal errors, we continue the accept loop
+            return ControlFlow::Continue(());
+        }
+    }
+
+    // If it's not a non-fatal I/O error, then treat it as fatal error
+    // This includes all other error types and I/O errors not explicitly handled above
+    ControlFlow::Break(e)
+}
 
 /// A type alias for the source of an error, which is a boxed trait object.
 /// This allows for dynamic dispatch and type erasure of the original error type.
